Removed legacy LB controller and replaced with the v1/Auth0 LB controller. (#406)
