+++ packages/backend/src/App.ts
@@ -24,10 +24,7 @@ app.use(morgan('dev'))
 
 app.use(
   cors({
-    origin: [
-      /.*-pocket-foundation\.vercel\.app/,
-      ...ALLOWED_DOMAINS,
-    ],
+    origin: [/.*-pocket-foundation\.vercel\.app/, ...ALLOWED_DOMAINS],
     methods: ['POST', 'PUT', 'GET', 'OPTIONS', 'HEAD'],
     credentials: true,
     exposedHeaders: ['Authorization'],
@@ -45,16 +42,6 @@ export const startServer = async (): Promise<void> => {
     app.listen(PORT, () => {
       console.log(`App listening to ${PORT}....`)
       console.log('Press Ctrl+C to quit.')
-
-      console.debug('DEBUG - TESTING JWT CHECK OBJECT', {
-        cache: true,
-        rateLimit: true,
-        jwksRequestsPerMinute: 5,
-        jwksUri: env('AUTH0_JWKS_URI'),
-        audience: env('AUTH0_AUDIENCE'),
-        issuer: env('AUTH0_ISSUER'),
-        algorithms: ['RS256'],
-      })
     })
   } catch (err) {
     console.error(err)
